Introduce terminate-after flag, graceful shutdown
This introduces a `--terminate-after` flag, which instructs a server to
gracefully shutdown after a specified number of requests.

The atomic counter logic lives in RequestHandler. When RequestHandler
detects the terminate limit has been hit, it sends a message to main,
which in turn shuts down all the servers and exits.

Signed-off-by: Andrew Seigner <siggy@buoyant.io>